Todos felices y contentos Validable no existe mas!!!!! yay!!! fin del problema

+++ awesome-tim/awesome-tim/EscalasDeEjes.cpp
@@ -23,12 +23,12 @@ double EscalasDeEjes::getCantidadUnidadesLogicasY(Sint32 pixelY)
 
 int EscalasDeEjes::getCantidadUnidadesFisicasX(double posX)
 {
-	return posX/escalaX;
+	return int(posX/escalaX);
 }
 
 int EscalasDeEjes::getCantidadUnidadesFisicasY(double posY)
 {
-	return posY/escalaY;
+	return int(posY/escalaY);
 }
 
 void EscalasDeEjes::setEscalaX(double relX){
